fix(datastore): make GetObjectLatestVersionNumber return uint32

+++ datastore/database/get_object_latest_version_number.go
@@ -8,8 +8,8 @@ import (
 	common_globals "github.com/PretendoNetwork/nex-protocols-common-go/v2/globals"
 )
 
-func GetObjectLatestVersionNumber(manager *common_globals.DataStoreManager, dataID types.UInt64) (uint16, *nex.Error) {
-	var version uint16
+func GetObjectLatestVersionNumber(manager *common_globals.DataStoreManager, dataID types.UInt64) (uint32, *nex.Error) {
+	var version uint32
 
 	err := manager.Database.QueryRow(`SELECT version FROM datastore.objects WHERE data_id=$1`, dataID).Scan(&version)
 	if err != nil {
